+++ src/main/java/org/jabref/logic/integrity/PersonNamesChecker.java
@@ -4,6 +4,7 @@ import java.util.Locale;
 import java.util.Optional;
 
 import org.jabref.logic.l10n.Localization;
+import org.jabref.model.entry.AuthorList;
 
 public class PersonNamesChecker implements ValueChecker {
 
@@ -15,6 +16,16 @@ public class PersonNamesChecker implements ValueChecker {
         } else if (valueTrimmedAndLowerCase.endsWith(" and") || valueTrimmedAndLowerCase.endsWith(",")) {
             return Optional.of(Localization.lang("should end with a name"));
         }
+
+        // Check that the value is in one of the two standard BibTeX formats:
+        //  Last, First and ...
+        //  First Last and ...
+        AuthorList authorList = AuthorList.parse(value);
+        if (!authorList.getAsLastFirstNamesWithAnd(false).equals(value)
+                && !authorList.getAsFirstLastNamesWithAnd().equals(value)) {
+            return Optional.of(Localization.lang("Names are not in the standard BibTeX format."));
+        }
+
         return Optional.empty();
     }
 }